@@ -1,39 +1,36 @@
-// import { SignedIn, SignedOut, SignInButton, UserButton } from '@clerk/clerk-react'
-// import { Button } from '@mui/material'
-// import { axiosInstance } from './lib/axios'
-import { Route, Routes } from 'react-router-dom';
-import HomePage from './pages/home/HomePage';
-import AuthCallbackPage from './pages/auth-callback/AuthCallbackPage';
-import { AuthenticateWithRedirectCallback } from '@clerk/clerk-react';
-import MainLayout from './layout/MainLayout';
-import ChatPage from './pages/chat/ChatPage';
-import AlbumPage from './pages/album/AlbumPage';
+import { Route, Routes } from "react-router-dom";
+import HomePage from "./pages/home/HomePage";
+import AuthCallbackPage from "./pages/auth-callback/AuthCallbackPage";
+import { AuthenticateWithRedirectCallback } from "@clerk/clerk-react";
+import MainLayout from "./layout/MainLayout";
+import ChatPage from "./pages/chat/ChatPage";
+import AlbumPage from "./pages/album/AlbumPage";
+import AdminPage from "./pages/admin/AdminPage";
 
-export default function App() {
-  // token
-  // const getSomeData = async() => {
-  //   await axiosInstance.get("/users",{
-  //     headers:{
-  //       "Authorization": `Bearer ${token}`
-  //     }
-  //   });
-  //   console.log(res)
-  // }
-  return (
-    <>
-      <Routes>
-        <Route
-          path='/sso-callback'
-          element={<AuthenticateWithRedirectCallback signUpForceRedirectUrl={"/auth-callback"} />}
-        />
-        <Route path="/auth-callback" element={<AuthCallbackPage />} />
+import { Toaster } from "react-hot-toast";
+import NotFoundPage from "./pages/404/NotFoundPage";
 
-        <Route element={<MainLayout />}>
-          <Route path="/" element={<HomePage />} />
-          <Route path="/chat" element={<ChatPage />} />
-          <Route path="/albums/:x" element={<AlbumPage />} />
-        </Route>
-      </Routes>
-    </>
-  )
+function App() {
+	return (
+		<>
+			<Routes>
+				<Route
+					path='/sso-callback'
+					element={<AuthenticateWithRedirectCallback signUpForceRedirectUrl={"/auth-callback"} />}
+				/>
+				<Route path='/auth-callback' element={<AuthCallbackPage />} />
+				<Route path='/admin' element={<AdminPage />} />
+
+				<Route element={<MainLayout />}>
+					<Route path='/' element={<HomePage />} />
+					<Route path='/chat' element={<ChatPage />} />
+					<Route path='/albums/:albumId' element={<AlbumPage />} />
+					<Route path='*' element={<NotFoundPage />} />
+				</Route>
+			</Routes>
+			<Toaster />
+		</>
+	);
 }
+
+export default App;
