@@ -1,36 +1,54 @@
-import { Container } from "@mui/material";
-import { Outlet } from "react-router-dom"
+import { ResizableHandle, ResizablePanel, ResizablePanelGroup } from "@/components/ui/resizable";
+import { Outlet } from "react-router-dom";
 import LeftSidebar from "./components/LeftSidebar";
+import FriendsActivity from "./components/FriendsActivity";
+import AudioPlayer from "./components/AudioPlayer";
+import { PlaybackControls } from "./components/PlaybackControls";
+import { useEffect, useState } from "react";
 
 const MainLayout = () => {
-  // const isMobile = false;
-  return (
-    <div className="h-screen bg-black text-white flex flex-col">
-      <Container  className="flex-1 flex h-full overflow-hidden p-2">
-        {/* left sidebar */}
-        <Container maxWidth="sm">
-          <LeftSidebar />
-        </Container>
+	const [isMobile, setIsMobile] = useState(false);
 
-        <Container maxWidth="sm" className="w-2 bg-black rounded-lg transition-colors">
-        </Container>
+	useEffect(() => {
+		const checkMobile = () => {
+			setIsMobile(window.innerWidth < 768);
+		};
 
-        {/* main content */}
-        <Container maxWidth="sm">
-          <Outlet />
-        </Container>
-        
-        <Container maxWidth="sm" className="w-2 bg-black rounded-lg transition-colors">
-        </Container>
+		checkMobile();
+		window.addEventListener("resize", checkMobile);
+		return () => window.removeEventListener("resize", checkMobile);
+	}, []);
 
-        {/* right sidebar */}
-        <Container maxWidth="sm">
-          friends activity 
-        </Container>
-        {/* Main Content */}
-      </Container>
-    </div>
-  )
-}
+	return (
+		<div className='h-screen bg-black text-white flex flex-col'>
+			<ResizablePanelGroup direction='horizontal' className='flex-1 flex h-full overflow-hidden p-2'>
+				<AudioPlayer />
+				{/* left sidebar */}
+				<ResizablePanel defaultSize={20} minSize={isMobile ? 0 : 10} maxSize={30}>
+					<LeftSidebar />
+				</ResizablePanel>
 
-export default MainLayout
+				<ResizableHandle className='w-2 bg-black rounded-lg transition-colors' />
+
+				{/* Main content */}
+				<ResizablePanel defaultSize={isMobile ? 80 : 60}>
+					<Outlet />
+				</ResizablePanel>
+
+				{!isMobile && (
+					<>
+						<ResizableHandle className='w-2 bg-black rounded-lg transition-colors' />
+
+						{/* right sidebar */}
+						<ResizablePanel defaultSize={20} minSize={0} maxSize={25} collapsedSize={0}>
+							<FriendsActivity />
+						</ResizablePanel>
+					</>
+				)}
+			</ResizablePanelGroup>
+
+			<PlaybackControls />
+		</div>
+	);
+};
+export default MainLayout;
