@@ -1,6 +1,6 @@
-import { Button } from "@mui/material";
-import { Slider } from "@mui/material";
-import { usePlayerStore } from "../../stores/usePlayerStore";
+import { Button } from "@/components/ui/button";
+import { Slider } from "@/components/ui/slider";
+import { usePlayerStore } from "@/stores/usePlayerStore";
 import { Laptop2, ListMusic, Mic2, Pause, Play, Repeat, Shuffle, SkipBack, SkipForward, Volume1 } from "lucide-react";
 import { useEffect, useRef, useState } from "react";
 
