@@ -1,5 +1,5 @@
 import axios from "axios";
 
 export const axiosInstance = axios.create({
-    baseURL: "http//localhost:5000/api",
-})
+	baseURL: import.meta.env.MODE === "development" ? "http://localhost:5000/api" : "/api",
+});
