@@ -1,52 +1,52 @@
-import { useEffect } from 'react';
-import Topbar from '../../components/ui/Topbar'
-import { useMusicStore } from '../../stores/useMusicStore'
-import { Slide } from '@mui/material';
-import FeaturedSection from './components/FeaturedSection';
-import SectionGrid from './components/SectionGrid';
+import Topbar from "@/components/Topbar";
+import { useMusicStore } from "@/stores/useMusicStore";
+import { useEffect } from "react";
+import FeaturedSection from "./components/FeaturedSection";
+import { ScrollArea } from "@/components/ui/scroll-area";
+import SectionGrid from "./components/SectionGrid";
+import { usePlayerStore } from "@/stores/usePlayerStore";
 
 const HomePage = () => {
-  const {
-    fetchFeaturedSongs,
-    fetchMadeForYouSongs,
-    fetchTrendingSongs,
-    isLoading,
-    madeForYouSongs,
-    featuredSongs,
-    trendingSongs,
-  } = useMusicStore();
+	const {
+		fetchFeaturedSongs,
+		fetchMadeForYouSongs,
+		fetchTrendingSongs,
+		isLoading,
+		madeForYouSongs,
+		featuredSongs,
+		trendingSongs,
+	} = useMusicStore();
 
-  useEffect(() => {
-    fetchFeaturedSongs();
-    fetchMadeForYouSongs();
-    fetchTrendingSongs();
-  },[fetchFeaturedSongs,fetchMadeForYouSongs,fetchTrendingSongs]);
+	const { initializeQueue } = usePlayerStore();
 
-  console.log({isLoading,madeForYouSongs,featuredSongs,trendingSongs});
-  
-  return (
-    <main 
-      className="rounded-md overflow-hidden h-full bg-gradient-to-b from-zinc-900 to-zinc-800"
-    >
-      <Topbar />
-      <Slide
-        className="h-[calc(100vh-180px]"
-      >
-        <div className="p-4 sm:p-6">
-          <h1 className="text-2xl sm:text-3xl font-bold mb-6">
-            Good afternoon
-          </h1>
-          <FeaturedSection />
-        <div className="space-y-8">
-          <p>made for you </p>
-          <p>trending</p>
-          <SectionGrid title="Made For You" songs={madeForYouSongs} isLoading={isLoading} />
-          <SectionGrid title="Trending" songs={trendingSongs} isLoading={isLoading} />
-        </div>
-        </div>
-      </Slide>
-    </main>
-  )
-}
+	useEffect(() => {
+		fetchFeaturedSongs();
+		fetchMadeForYouSongs();
+		fetchTrendingSongs();
+	}, [fetchFeaturedSongs, fetchMadeForYouSongs, fetchTrendingSongs]);
 
-export default HomePage
+	useEffect(() => {
+		if (madeForYouSongs.length > 0 && featuredSongs.length > 0 && trendingSongs.length > 0) {
+			const allSongs = [...featuredSongs, ...madeForYouSongs, ...trendingSongs];
+			initializeQueue(allSongs);
+		}
+	}, [initializeQueue, madeForYouSongs, trendingSongs, featuredSongs]);
+
+	return (
+		<main className='rounded-md overflow-hidden h-full bg-gradient-to-b from-zinc-800 to-zinc-900'>
+			<Topbar />
+			<ScrollArea className='h-[calc(100vh-180px)]'>
+				<div className='p-4 sm:p-6'>
+					<h1 className='text-2xl sm:text-3xl font-bold mb-6'>Good afternoon</h1>
+					<FeaturedSection />
+
+					<div className='space-y-8'>
+						<SectionGrid title='Made For You' songs={madeForYouSongs} isLoading={isLoading} />
+						<SectionGrid title='Trending' songs={trendingSongs} isLoading={isLoading} />
+					</div>
+				</div>
+			</ScrollArea>
+		</main>
+	);
+};
+export default HomePage;
