@@ -1,50 +1,48 @@
-import { Button } from "@mui/material";
-import { Song } from "../../../types";
+import { Song } from "@/types";
 import SectionGridSkeleton from "./SectionGridSkeleton";
+import { Button } from "@/components/ui/button";
+import PlayButton from "./PlayButton";
 
 type SectionGridProps = {
-  title:string;
-  songs:Song[];
-  isLoading:boolean;
-}
-const SectionGrid = ({songs,title,isLoading}:SectionGridProps) => {
-  if(isLoading) return <SectionGridSkeleton />
-  return (
-    <div className="mb-8">
-      <div className="flex items-center justify-between mb-4">
-        <h2 className="text-xl sm:text-2xl font-bold">title</h2>
-        <Button className="text-sm text-zinc-400 hover:text-white">
-          Show all
-        </Button>
-      </div>
+	title: string;
+	songs: Song[];
+	isLoading: boolean;
+};
+const SectionGrid = ({ songs, title, isLoading }: SectionGridProps) => {
+	if (isLoading) return <SectionGridSkeleton />;
 
-      <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-4 gap-4">
-        {songs.map((song) => (
-          <div
-            key={song._id}
-            className="bg-zinc-800/40 p-4 rounded-md hover:bg-zinc-700/40 transition-all group cursor-pointer"
-          >
-            <div className="relative mb-4">
-              <div className="aspect-square">
-                <img 
-                  src={song.imageUrl}
-                  alt={song.title}
-                  className="w-full h-full obkect-cover transition-transform duration-300 group-hover:scale-105" 
-                />
-              </div>
-            </div>
-            <h3 className="font-medium mb-2 truncate">
-              {song.title}
-            </h3>
-            <p className="text-sm text-zinc-400 truncate">
-              {song.artist}
-            </p>
-          </div>
+	return (
+		<div className='mb-8'>
+			<div className='flex items-center justify-between mb-4'>
+				<h2 className='text-xl sm:text-2xl font-bold'>{title}</h2>
+				<Button variant='link' className='text-sm text-zinc-400 hover:text-white'>
+					Show all
+				</Button>
+			</div>
 
-        ))}
-      </div>
-    </div>
-  )
-}
-
-export default SectionGrid
+			<div className='grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-4 gap-4'>
+				{songs.map((song) => (
+					<div
+						key={song._id}
+						className='bg-zinc-800/40 p-4 rounded-md hover:bg-zinc-700/40 transition-all group cursor-pointer'
+					>
+						<div className='relative mb-4'>
+							<div className='aspect-square rounded-md shadow-lg overflow-hidden'>
+								<img
+									src={song.imageUrl}
+									alt={song.title}
+									className='w-full h-full object-cover transition-transform duration-300 
+									group-hover:scale-105'
+								/>
+							</div>
+							<PlayButton song={song} />
+						</div>
+						<h3 className='font-medium mb-2 truncate'>{song.title}</h3>
+						<p className='text-sm text-zinc-400 truncate'>{song.artist}</p>
+					</div>
+				))}
+			</div>
+		</div>
+	);
+};
+export default SectionGrid;
