@@ -1,44 +1,52 @@
-import { useEffect, useState } from 'react'
-import { axiosInstance } from '../lib/axios';
-import { useAuth } from '@clerk/clerk-react';
-import { Loader } from 'lucide-react';
-import { useAuthStore } from '../stores/useAuthStore';
+import { axiosInstance } from "@/lib/axios";
+import { useAuthStore } from "@/stores/useAuthStore";
+import { useChatStore } from "@/stores/useChatStore";
+import { useAuth } from "@clerk/clerk-react";
+import { Loader } from "lucide-react";
+import { useEffect, useState } from "react";
 
 const updateApiToken = (token: string | null) => {
-  if (token) axiosInstance.defaults.headers.common['Authorization'] = `Bearer ${token}`;
-  else delete axiosInstance.defaults.headers.common["Authorization"];
+	if (token) axiosInstance.defaults.headers.common["Authorization"] = `Bearer ${token}`;
+	else delete axiosInstance.defaults.headers.common["Authorization"];
 };
 
 const AuthProvider = ({ children }: { children: React.ReactNode }) => {
-  const { getToken } = useAuth()
-  const [loading, setLoading] = useState(true)
-  const { checkAdminStatus } = useAuthStore();
-  useEffect(() => {
-    const initAuth = async () => {
-      try {
-        const token = await getToken();
-        updateApiToken(token);
-        if(token) {
-          await checkAdminStatus();
-        }
-      } catch (error: any) {
-        updateApiToken(null);
-        console.log("Error in auth provider", error);
-      } finally {
-        setLoading(false);
-      }
-    };
-    initAuth();
-  }, [getToken]);
+	const { getToken, userId } = useAuth();
+	const [loading, setLoading] = useState(true);
+	const { checkAdminStatus } = useAuthStore();
+	const { initSocket, disconnectSocket } = useChatStore();
 
-  if (loading) return (
-    <div className="h-screen w-full flex items-center justify-center">
-      <Loader className="size-8 text-emerald-500 animate-spin" />
-    </div>
-  )
-  return (
-    <div>{children}</div>
-  )
-}
+	useEffect(() => {
+		const initAuth = async () => {
+			try {
+				const token = await getToken();
+				updateApiToken(token);
+				if (token) {
+					await checkAdminStatus();
+					// init socket
+					if (userId) initSocket(userId);
+				}
+			} catch (error: any) {
+				updateApiToken(null);
+				console.log("Error in auth provider", error);
+			} finally {
+				setLoading(false);
+			}
+		};
 
-export default AuthProvider
+		initAuth();
+
+		// clean up
+		return () => disconnectSocket();
+	}, [getToken, userId, checkAdminStatus, initSocket, disconnectSocket]);
+
+	if (loading)
+		return (
+			<div className='h-screen w-full flex items-center justify-center'>
+				<Loader className='size-8 text-emerald-500 animate-spin' />
+			</div>
+		);
+
+	return <>{children}</>;
+};
+export default AuthProvider;
