@@ -1,33 +1,33 @@
-import {create} from "zustand";
-import { axiosInstance } from "../lib/axios";
+import { axiosInstance } from "@/lib/axios";
+import { create } from "zustand";
 
 interface AuthStore {
-  isAdmin : boolean;
-  isLoading: boolean;
-  error: string | null;
+	isAdmin: boolean;
+	isLoading: boolean;
+	error: string | null;
 
-  checkAdminStatus: () => Promise<void>;
-  reset: () => void;
+	checkAdminStatus: () => Promise<void>;
+	reset: () => void;
 }
 
 export const useAuthStore = create<AuthStore>((set) => ({
-  isAdmin: false,
-  isLoading: false,
-  error: null,
+	isAdmin: false,
+	isLoading: false,
+	error: null,
 
-  checkAdminStatus: async () => {
-    set({ isLoading:true,error:null });
-    try {
-        const response = await axiosInstance.get("/admin/check");
-        set({isAdmin:response.data.admin})
-    } catch (error: any) {
-      set({ isAdmin:false,error:error.response.data.message });
-    } finally {
-      set({ isLoading: false });
-    }
-  },
+	checkAdminStatus: async () => {
+		set({ isLoading: true, error: null });
+		try {
+			const response = await axiosInstance.get("/admin/check");
+			set({ isAdmin: response.data.admin });
+		} catch (error: any) {
+			set({ isAdmin: false, error: error.response.data.message });
+		} finally {
+			set({ isLoading: false });
+		}
+	},
 
-  reset: () => {
-    set({ isAdmin: false, isLoading: false, error: null});
-  }
-}))
+	reset: () => {
+		set({ isAdmin: false, isLoading: false, error: null });
+	},
+}));
