@@ -1,5 +1,5 @@
 import { create } from "zustand";
-import { Song } from "../types";
+import { Song } from "@/types";
 import { useChatStore } from "./useChatStore";
 
 interface PlayerStore {
@@ -35,13 +35,13 @@ export const usePlayerStore = create<PlayerStore>((set, get) => ({
 
 		const song = songs[startIndex];
 
-		// const socket = useChatStore.getState().socket;
-		// if (socket.auth) {
-		// 	socket.emit("update_activity", {
-		// 		userId: socket.auth.userId,
-		// 		activity: `Playing ${song.title} by ${song.artist}`,
-		// 	});
-		// }
+		const socket = useChatStore.getState().socket;
+		if (socket.auth) {
+			socket.emit("update_activity", {
+				userId: socket.auth.userId,
+				activity: `Playing ${song.title} by ${song.artist}`,
+			});
+		}
 		set({
 			queue: songs,
 			currentSong: song,
@@ -53,13 +53,13 @@ export const usePlayerStore = create<PlayerStore>((set, get) => ({
 	setCurrentSong: (song: Song | null) => {
 		if (!song) return;
 
-		// const socket = useChatStore.getState().socket;
-		// if (socket.auth) {
-		// 	socket.emit("update_activity", {
-		// 		userId: socket.auth.userId,
-		// 		activity: `Playing ${song.title} by ${song.artist}`,
-		// 	});
-		// }
+		const socket = useChatStore.getState().socket;
+		if (socket.auth) {
+			socket.emit("update_activity", {
+				userId: socket.auth.userId,
+				activity: `Playing ${song.title} by ${song.artist}`,
+			});
+		}
 
 		const songIndex = get().queue.findIndex((s) => s._id === song._id);
 		set({
@@ -73,14 +73,14 @@ export const usePlayerStore = create<PlayerStore>((set, get) => ({
 		const willStartPlaying = !get().isPlaying;
 
 		const currentSong = get().currentSong;
-		// const socket = useChatStore.getState().socket;
-		// if (socket.auth) {
-		// 	socket.emit("update_activity", {
-		// 		userId: socket.auth.userId,
-		// 		activity:
-		// 			willStartPlaying && currentSong ? `Playing ${currentSong.title} by ${currentSong.artist}` : "Idle",
-		// 	});
-		// }
+		const socket = useChatStore.getState().socket;
+		if (socket.auth) {
+			socket.emit("update_activity", {
+				userId: socket.auth.userId,
+				activity:
+					willStartPlaying && currentSong ? `Playing ${currentSong.title} by ${currentSong.artist}` : "Idle",
+			});
+		}
 
 		set({
 			isPlaying: willStartPlaying,
@@ -95,13 +95,13 @@ export const usePlayerStore = create<PlayerStore>((set, get) => ({
 		if (nextIndex < queue.length) {
 			const nextSong = queue[nextIndex];
 
-			// const socket = useChatStore.getState().socket;
-			// if (socket.auth) {
-			// 	socket.emit("update_activity", {
-			// 		userId: socket.auth.userId,
-			// 		activity: `Playing ${nextSong.title} by ${nextSong.artist}`,
-			// 	});
-			// }
+			const socket = useChatStore.getState().socket;
+			if (socket.auth) {
+				socket.emit("update_activity", {
+					userId: socket.auth.userId,
+					activity: `Playing ${nextSong.title} by ${nextSong.artist}`,
+				});
+			}
 
 			set({
 				currentSong: nextSong,
@@ -112,13 +112,13 @@ export const usePlayerStore = create<PlayerStore>((set, get) => ({
 			// no next song
 			set({ isPlaying: false });
 
-			// const socket = useChatStore.getState().socket;
-			// if (socket.auth) {
-			// 	socket.emit("update_activity", {
-			// 		userId: socket.auth.userId,
-			// 		activity: `Idle`,
-			// 	});
-			// }
+			const socket = useChatStore.getState().socket;
+			if (socket.auth) {
+				socket.emit("update_activity", {
+					userId: socket.auth.userId,
+					activity: `Idle`,
+				});
+			}
 		}
 	},
 	playPrevious: () => {
@@ -129,13 +129,13 @@ export const usePlayerStore = create<PlayerStore>((set, get) => ({
 		if (prevIndex >= 0) {
 			const prevSong = queue[prevIndex];
 
-			// const socket = useChatStore.getState().socket;
-			// if (socket.auth) {
-			// 	socket.emit("update_activity", {
-			// 		userId: socket.auth.userId,
-			// 		activity: `Playing ${prevSong.title} by ${prevSong.artist}`,
-			// 	});
-			// }
+			const socket = useChatStore.getState().socket;
+			if (socket.auth) {
+				socket.emit("update_activity", {
+					userId: socket.auth.userId,
+					activity: `Playing ${prevSong.title} by ${prevSong.artist}`,
+				});
+			}
 
 			set({
 				currentSong: prevSong,
@@ -146,13 +146,13 @@ export const usePlayerStore = create<PlayerStore>((set, get) => ({
 			// no prev song
 			set({ isPlaying: false });
 
-			// const socket = useChatStore.getState().socket;
-			// if (socket.auth) {
-			// 	socket.emit("update_activity", {
-			// 		userId: socket.auth.userId,
-			// 		activity: `Idle`,
-			// 	});
-			// }
+			const socket = useChatStore.getState().socket;
+			if (socket.auth) {
+				socket.emit("update_activity", {
+					userId: socket.auth.userId,
+					activity: `Idle`,
+				});
+			}
 		}
 	},
 }));
